extract cdc scan internal structs
