+++ ydb/core/tx/datashard/ya.make
@@ -206,6 +206,7 @@ SRCS(
     store_scheme_tx_unit.cpp
     store_snapshot_tx_unit.cpp
     store_write_unit.cpp
+    stream_scan_common.cpp
     upload_stats.cpp
     volatile_tx.cpp
     wait_for_plan_unit.cpp
